pipe/phono-deps: move component interfaces from phono to pipe
