+++ pipe/runner.go
@@ -2,13 +2,11 @@ package pipe
 
 import (
 	"io"
-
-	"github.com/pipelined/phono"
 )
 
 // pumpRunner is pump's runner.
 type pumpRunner struct {
-	phono.Pump
+	Pump
 	fn  func() ([][]float64, error)
 	out chan message
 	hooks
@@ -16,7 +14,7 @@ type pumpRunner struct {
 
 // processRunner represents processor's runner.
 type processRunner struct {
-	phono.Processor
+	Processor
 	fn  func([][]float64) ([][]float64, error)
 	in  <-chan message
 	out chan message
@@ -25,7 +23,7 @@ type processRunner struct {
 
 // sinkRunner represents sink's runner.
 type sinkRunner struct {
-	phono.Sink
+	Sink
 	fn func([][]float64) error
 	in <-chan message
 	hooks
@@ -93,7 +91,7 @@ func resetter(i interface{}) hook {
 
 // newPumpRunner creates the closure. it's separated from run to have pre-run
 // logic executed in correct order for all components.
-func newPumpRunner(pipeID string, bufferSize int, p phono.Pump) (*pumpRunner, int, int, error) {
+func newPumpRunner(pipeID string, bufferSize int, p Pump) (*pumpRunner, int, int, error) {
 	fn, sampleRate, numChannels, err := p.Pump(pipeID, bufferSize)
 	if err != nil {
 		return nil, 0, 0, err
@@ -171,7 +169,7 @@ func (r *pumpRunner) run(pipeID, componentID string, cancel <-chan struct{}, pro
 
 // newProcessRunner creates the closure. it's separated from run to have pre-run
 // logic executed in correct order for all components.
-func newProcessRunner(pipeID string, sampleRate, numChannels, bufferSize int, p phono.Processor) (*processRunner, error) {
+func newProcessRunner(pipeID string, sampleRate, numChannels, bufferSize int, p Processor) (*processRunner, error) {
 	fn, err := p.Process(pipeID, sampleRate, numChannels, bufferSize)
 	if err != nil {
 		return nil, err
@@ -234,7 +232,7 @@ func (r *processRunner) run(pipeID, componentID string, cancel chan struct{}, in
 
 // newSinkRunner creates the closure. it's separated from run to have pre-run
 // logic executed in correct order for all components.
-func newSinkRunner(pipeID string, sampleRate, numChannels, bufferSize int, s phono.Sink) (*sinkRunner, error) {
+func newSinkRunner(pipeID string, sampleRate, numChannels, bufferSize int, s Sink) (*sinkRunner, error) {
 	fn, err := s.Sink(pipeID, sampleRate, numChannels, bufferSize)
 	if err != nil {
 		return nil, err
